# docs/api.yaml
openapi: 3.0.3
info:
  title: URL Shortener - Shorten & Redirect Core API
  version: "1.0.0"
  description: >
    Core API for creating short links from long URLs, resolving them via redirect, and reporting
    click statistics. Statistics are a single running total per link (count + last-accessed time),
    not a time-series or per-click event log. A short link may optionally be given an expiration;
    an expired code behaves identically to one that never existed — same 404, no distinguishing
    signal.
servers:
  - url: /
paths:
  /api/links:
    post:
      operationId: createShortLink
      summary: Create a short link for a long URL
      description: >
        Creates a new short link for the given destination URL. If the destination URL
        (exact string match, after whitespace trimming) already has a short link, returns
        the existing one instead of creating a duplicate.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateLinkRequest'
      responses:
        '201':
          description: Short link created (or an existing matching one returned)
          headers:
            Location:
              description: The fully-qualified short URL
              schema:
                type: string
                format: uri
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LinkResponse'
        '400':
          description: The submitted URL failed validation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '429':
          description: Too many creation requests from this caller
          headers:
            Retry-After:
              description: Seconds to wait before retrying
              schema:
                type: integer
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /{code}:
    get:
      operationId: redirectToLongUrl
      summary: Resolve a short code and redirect to its long URL
      parameters:
        - name: code
          in: path
          required: true
          description: 6-character short code (case-insensitive)
          schema:
            type: string
      responses:
        '302':
          description: Redirect to the long URL associated with this short code
          headers:
            Location:
              description: The original long URL
              schema:
                type: string
                format: uri
        '404':
          description: No short link exists for this code
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/links/{code}/stats:
    get:
      operationId: getStats
      summary: Get click statistics for a short link
      parameters:
        - name: code
          in: path
          required: true
          description: 6-character short code (case-insensitive)
          schema:
            type: string
      responses:
        '200':
          description: Current click count and access time for this short code
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LinkStatsResponse'
        '404':
          description: No short link exists for this code
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '429':
          description: Too many stats requests from this caller
          headers:
            Retry-After:
              description: Seconds to wait before retrying
              schema:
                type: integer
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
components:
  schemas:
    CreateLinkRequest:
      type: object
      required: [url]
      properties:
        url:
          type: string
          format: uri
          maxLength: 2048
          example: "https://example.com/some/very/long/path?with=query"
        expiresInSeconds:
          type: integer
          format: int64
          minimum: 1
          maximum: 31536000
          nullable: true
          description: >
            Optional. How long the short link should remain resolvable. Omit for a link that
            never expires (the default — expiry is opt-in, not automatic).
          example: 3600
    LinkResponse:
      type: object
      properties:
        shortCode:
          type: string
          example: "a1b2c3"
        shortUrl:
          type: string
          format: uri
          example: "https://short.example/a1b2c3"
        longUrl:
          type: string
          format: uri
          example: "https://example.com/some/very/long/path?with=query"
        createdAt:
          type: string
          format: date-time
        expiresAt:
          type: string
          format: date-time
          nullable: true
          description: Null if the link was created with no expiration.
    LinkStatsResponse:
      type: object
      properties:
        shortCode:
          type: string
          example: "a1b2c3"
        longUrl:
          type: string
          format: uri
          example: "https://example.com/some/very/long/path?with=query"
        clickCount:
          type: integer
          format: int64
          example: 42
        createdAt:
          type: string
          format: date-time
        lastAccessedAt:
          type: string
          format: date-time
          nullable: true
          description: Null if the link has never been visited
        expiresAt:
          type: string
          format: date-time
          nullable: true
          description: Null if the link was created with no expiration.
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          description: Machine-readable error code
          example: "VALIDATION_ERROR"
        message:
          type: string
          description: Human-readable explanation
          example: "URL scheme must be http or https"
        timestamp:
          type: string
          format: date-time